Move exceptions-related API to the correct header file.
